requirements.txt
playwright
apify-client
